Implicit cvar copy fixed
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@8063 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ framework/Session.cpp
@@ -2889,7 +2889,7 @@ void idSessionLocal::Frame() {
 		r_swapIntervalTemp = 1;
 	} else
 		if ( r_swapIntervalTemp ^ r_swapInterval ) // avoid useless hassle
-			r_swapIntervalTemp = r_swapInterval;
+			r_swapIntervalTemp = (int)r_swapInterval;
 
 	if ( !mapSpawned ) {
 		return;
